change font

+++ experiments/test_nag.py
@@ -24,7 +24,7 @@ with open("../prompts/test_prompts.json.new", "r") as f:
 
 def run(nag_scale, nag_alpha, nag_tau):
     wandb.init(project="nag-sweep")
-    score = np.array([0, 0, 0], dtype=int)
+    score = np.array([0, 0, 0], dtype=float)
     total = 0
     for seed in range(2):
         for i in dev_prompts:
